+++ CHANGELOG.md
@@ -723,3 +723,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix ZHS, ZHT, and JPN not displaying color in tooltips (Celicath)
 
 #### dev ####
+* Card descriptors (kiooeht)